This checkin represents some cleanup and refactoring.
The message queue table has been rename from 'test' to 'mq'
Here is the DDL:

CREATE TABLE IF NOT EXISTS mq
( id INTEGER PRIMARY KEY,
contractAddress VARCHAR(50) NOT NULL,
ballotId INTEGER NOT NULL,
toEmail VARCHAR(125) NOT NULL,
block INTEGER NOT NULL,
memo TEXT,
status VARCHAR(25) NOT NULL DEFAULT 'initial',
created DATETIME NOT NULL DEFAULT ( Datetime('now') ),
updated DATETIME, \
unique ( contractAddress, ballotId, toEmail )
);

Things are run as services using node-monitor.

When queuing rows in mq it will be one for each address in the email.yaml file.  This is denormalized but easier to track.
The email "template" is embedded in queue.js which less that ideal.

Function points have been exercised
When queuing rows in mq it will be one for each address in the email.yaml file.  This is denormalized but easier to track.
The email "template" is embedded in queue.js which less that ideal.

Function points have been exercised.